requirements.txt
pandas
docx
snowflake-connector-python
snowflake-snowpark-python
cryptography
toml
